Small refactor following project patterns
